move axes into their own component
